requirements.txt
pymysql
requests
python-dotenv
cryptography
